Drop binarization, feed continuous grayscale into detectors
ImageLoader used to collapse the 8-bit gray channel to ±1 via a hard
threshold, discarding sub-pixel antialiasing information on stroke
boundaries. Normalize the channel to [-1.0, +1.0] instead (0 → -1,
255 → +1) and propagate continuous values through RawImage,
DetectorMask, DetectorCluster, and ImageEncoder.

Prototype data is still defined in ±1 (the ideal binary limit); convert
to DoubleArray at config-generation time so prototype codes remain
exactly what they were. Activation threshold stays at 5.0, where ideal
prototypes produce clean single-bit-per-detector codes and real
MNIST pixels (center ≥ 0.66 after normalization) still exceed it.

With the previous step-prototype pass eliminating staircase-corner
codes, grayscale input now produces zero unknown cluster codes on the
test images — the dictionary fully covers what the encoder emits.

+++ src/main/kotlin/perception/visual/data/RawImage.kt
@@ -1,13 +1,13 @@
 package org.example.perception.visual.data
 
 class RawImage(
-    private val pixels: Array<IntArray>,
+    private val pixels: Array<DoubleArray>,
 ) {
 
     val height: Int = pixels.size
     val width: Int = if (height > 0) pixels[0].size else 0
 
-    fun getPixel(x: Int, y: Int): Int {
+    fun getPixel(x: Int, y: Int): Double {
         return pixels[y][x]
     }
 }
